little cleaning. sometimes servres doesn't send response if the same data has been posted many times
