+++ Makefile
@@ -2,7 +2,7 @@ GNU = c++
 FLAGS = -Wall -Wextra -Werror
 C_98 = -std=c++98
 CPP_FILES = SocketServer.cpp server.cpp ServerParseRequest.cpp ServerResponse.cpp InfoServer.cpp \
-			ServerStatusCode.cpp
+			ServerStatusCode.cpp StringManipulations.cpp PrintingFunctions.cpp
 CPP_OBJ = $(CPP_FILES:.cpp=.o)
 NAME = server
