+++ StringManipulations.hpp
@@ -1,0 +1,10 @@
+#ifndef STRING_MANIPULATIONS_H
+#define STRING_MANIPULATIONS_H
+
+#include <iostream>
+#include <unistd.h>
+#include <cstring>
+
+void	*ft_memset(void *s, int c, size_t len);
+
+#endif
